Fetch testing

+++ src/components/SignIn.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { auth, provider } from "../firebase";
 import { signInWithPopup } from "firebase/auth";
-
+import { fetchAPI } from "../service/fetchapi";
 
 const SignIn = ({ setUser }) => {
     const handleClick = async () => {
@@ -11,14 +11,19 @@ const SignIn = ({ setUser }) => {
         const email = data.user.email;
         const username = data.user.displayName;
 
-        await fetch(`${import.meta.env.VITE_API_URL}/api/users`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ email, username, theme: "light" })
-        }); 
-
-        localStorage.setItem("email", email);
-        setUser(email);
+        const existingUser = await fetchAPI(`/users/${email}`, "GET");
+        
+        if(existingUser) {
+            localStorage.setItem("email", email);
+            setUser(email);
+        } else{
+            const newUser = await fetchAPI('/users', "POST", { email, username, theme: "dark" });
+            if (newUser) {
+                localStorage.setItem('email', email);
+                console.log("Log In Success");
+            }
+        }
+        
         console.log("Log In Success");
         } catch (err) {
         console.error("Login error:", err.message);
